Move resources to sub package
