+++ jabsrv/src/main/java/module-info.java
@@ -13,6 +13,8 @@ module org.jabref.jabsrv {
     opens org.jabref.http.server.command to com.google.gson, org.glassfish.hk2.locator, org.glassfish.hk2.utilities, com.fasterxml.jackson.databind;
     exports org.jabref.http.server.services;
     exports org.jabref.http;
+    opens org.jabref.http.server.resources to org.glassfish.hk2.locator, org.glassfish.hk2.utilities;
+    exports org.jabref.http.server.resources;
 
     requires javafx.base;
