+++ jabsrv/src/main/java/org/jabref/http/server/resources/MapResource.java
@@ -1,4 +1,4 @@
-package org.jabref.http.server;
+package org.jabref.http.server.resources;
 
 import java.io.IOException;
 import java.nio.file.Files;
