+++ jabsrv/src/test/java/org/jabref/http/server/LibraryResourceTest.java
@@ -1,6 +1,8 @@
 package org.jabref.http.server;
 
 import org.jabref.http.JabrefMediaType;
+import org.jabref.http.server.resources.LibrariesResource;
+import org.jabref.http.server.resources.LibraryResource;
 
 import jakarta.ws.rs.core.Application;
 import org.glassfish.jersey.server.ResourceConfig;
